Backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.3.3
chardet==5.2.0
colorama==0.4.6
coverage==7.5.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==22.0.0
iniconfig==2.0.0
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.1.19
psycopg-binary==3.1.19
PyJWT==2.8.0
pyproject-api==1.6.1
pytest==8.2.0
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.26.2
whitenoise==6.6.0
drf-spectacular==0.27.2
